+++ tool/lib/src/preparing_tests/commands.dart
@@ -11,20 +11,31 @@ class FlutterCommand {
     String target,
     String device, {
     @required String flavor,
+    String additionalArguments,
   }) {
     // ignore: missing_whitespace_between_adjacent_strings
     return 'flutter run -d $device --target=$target'
-        '${flavor != null ? ' --flavor $flavor' : ''}';
+        '${flavor != null ? ' --flavor $flavor' : ''}'
+        '${additionalArguments != null ? ' $additionalArguments' : ''}';
   }
 
   String attach(String debugUri, String device) =>
       'flutter attach -d $device --debug-uri $debugUri';
 
-  String dart(String file, [Map<String, String> arguments]) {
-    final args = arguments?.entries
+  String dart(
+    String file, {
+    Map<String, String> testArguments,
+    String dartArguments,
+  }) {
+    final args = testArguments?.entries
         ?.map((entry) =>
             '${entry.key}${entry.value.isNotEmpty ? ' ${entry.value}' : ''}')
         ?.join(' ');
-    return 'dart $file ${args ?? ''}';
+
+    return 'dart'
+        '${dartArguments != null ? ' $dartArguments' : ''}'
+        // ignore: missing_whitespace_between_adjacent_strings
+        ' $file'
+        '${args != null ? ' $args' : ''}';
   }
 }